# web/src/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Sard API",
    "license" : {
      "name" : "AGPL-3.0-only",
      "identifier" : "AGPL-3.0-only"
    },
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "security" : [ {
    "session" : [ ]
  } ],
...
  "paths" : {
...
    "/api/v1/self-backup/repository" : {
      "put" : {
        "tags" : [ "self-backup" ],
        "summary" : "Bind the self-backup to a repository of the built-in agent",
        "description" : "Creates the two system sources with a nightly schedule, or moves them to this repository and the current built-in agent; binding again unchanged changes nothing. 422 codes: self_agent_missing, unknown_repository, repository_not_initialized, local_storage_unconfirmed, unknown_plugin, invalid_config (the built-in agent lacks the database password's secret), validation_failed.",
        "operationId" : "bindSelfBackupRepository",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SelfBackupRepositoryInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "422" : {
            "description" : "Rejected values; errors name the fields",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SelfBackup"
                }
              }
            }
          },
          "401" : {
            "description" : "No session or it expired",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            }
          },
          "403" : {
            "description" : "Origin does not match the request (CSRF)",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            }
          },
          "503" : {
            "description" : "The database is unavailable; nothing was changed",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/self-backup/runs" : {
      "post" : {
        "tags" : [ "self-backup" ],
        "summary" : "Back up the installation now",
        "description" : "Queues a manual run of each system source; a source with an active run answers with that run. 409 self_backup_not_configured before a repository is bound; agent_revoked, unknown_plugin or unknown_repository when a run cannot start.",
        "operationId" : "startSelfBackup",
        "responses" : {
          "409" : {
            "description" : "Not configured, or a run cannot start (code)",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SelfBackupRuns"
                }
              }
            }
          },
          "401" : {
            "description" : "No session or it expired",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            }
          },
          "403" : {
            "description" : "Origin does not match the request (CSRF)",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            }
          },
          "503" : {
            "description" : "The database is unavailable; nothing was changed",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/self-backup" : {
      "get" : {
        "tags" : [ "self-backup" ],
        "summary" : "The self-backup's state",
        "operationId" : "getSelfBackup",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SelfBackup"
                }
              }
            }
          },
          "401" : {
            "description" : "No session or it expired",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            }
          },
          "503" : {
            "description" : "The database is unavailable; nothing was changed",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs" : {
      "get" : {
        "tags" : [ "runs" ],
        "summary" : "List runs",
        "description" : "Newest first. Filters combine with AND; status values with OR. agentId is the agent that ran the step, not the source's current agent.",
        "operationId" : "listRuns",
        "parameters" : [ {
          "name" : "sourceId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "agentId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RunStatus"
            }
          }
        }, {
          "name" : "queuedFrom",
          "in" : "query",
          "description" : "Queued at or after this time",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "queuedTo",
          "in" : "query",
          "description" : "Queued before this time; must not be earlier than queuedFrom",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "cursor",
          "in" : "query",
          "description" : "nextCursor of the previous page; omit for the first page",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "Page size",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50,
            "maximum" : 200,
            "minimum" : 1
          }
        } ],
        "responses" : {
          "422" : {
            "description" : "Rejected values; errors name the fields",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RunPage"
                }
              }
            }
          },
          "401" : {
            "description" : "No session or it expired",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            }
          },
          "503" : {
            "description" : "The database is unavailable; nothing was changed",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ErrorCode" : {
        "type" : "string",
        "enum" : [ "unauthenticated", "too_many_attempts", "not_found", "validation_failed", "unknown_agent", "unknown_plugin", "unknown_repository", "invalid_config", "run_active", "token_used", "token_expired", "not_implemented", "origin_rejected", "agent_revoked", "unavailable", "self_agent_confirmation_required", "system_source", "self_agent_missing", "repository_not_initialized", "local_storage_unconfirmed", "self_backup_not_configured" ]
      },
      "Problem" : {
        "type" : "object",
        "description" : "An error (RFC 9457)",
        "properties" : {
          "type" : {
            "type" : "string",
            "description" : "URI reference identifying the problem type; about:blank when the status says it all"
          },
          "title" : {
            "type" : "string",
            "description" : "Short summary for people; not for matching, use code"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP status code"
          },
          "detail" : {
            "type" : [ "string", "null" ],
            "description" : "Explanation of this occurrence; null when the title says it all"
          },
          "code" : {
            "$ref" : "#/components/schemas/ErrorCode"
          }
        },
        "required" : [ "code", "detail", "status", "title", "type" ]
      },
      "FieldError" : {
        "type" : "object",
        "description" : "A field of the request that failed validation",
        "properties" : {
          "field" : {
            "type" : "string",
            "description" : "JSON name of the field, e.g. repositoryName"
          },
          "message" : {
            "type" : "string",
            "description" : "Why the value was rejected"
          }
        },
        "required" : [ "field", "message" ]
      },
      "ValidationProblem" : {
        "type" : "object",
        "description" : "The request is well-formed but its values are not acceptable (RFC 9457)",
        "properties" : {
          "type" : {
            "type" : "string",
            "description" : "URI reference identifying the problem type; about:blank when the status says it all"
          },
          "title" : {
            "type" : "string",
            "description" : "Short summary for people; not for matching, use code"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP status code"
          },
          "detail" : {
            "type" : [ "string", "null" ],
            "description" : "Explanation of this occurrence; null when the title says it all"
          },
          "code" : {
            "$ref" : "#/components/schemas/ErrorCode"
          },
          "errors" : {
            "type" : "array",
            "description" : "The rejected fields",
            "items" : {
              "$ref" : "#/components/schemas/FieldError"
            }
          }
        },
        "required" : [ "code", "detail", "errors", "status", "title", "type" ]
      },
...
      "SystemSourceRole" : {
        "type" : "string",
        "enum" : [ "self_database", "self_keys" ]
      },
...
      "SelfBackupRepositoryInput" : {
        "type" : "object",
        "description" : "The repository of the built-in agent the self-backup goes to",
        "properties" : {
          "repositoryName" : {
            "type" : "string",
            "description" : "A repository of the built-in agent's last Register, initialised"
          },
          "confirmLocalStorage" : {
            "type" : "boolean",
            "description" : "Required true for a local repository: a copy on the server's own machine is lost with the machine"
          }
        },
        "required" : [ "repositoryName" ]
      },
      "SelfBackup" : {
        "type" : "object",
        "description" : "The self-backup of the installation: the database, the CA and the configuration",
        "properties" : {
          "configured" : {
            "type" : "boolean",
            "description" : "A repository is bound and both system sources exist"
          },
          "agentId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid",
            "description" : "The live built-in agent (sard-self); null without one"
          },
          "repository" : {
            "description" : "Null until configured",
            "oneOf" : [ {
              "$ref" : "#/components/schemas/SelfBackupRepository"
            }, {
              "type" : "null"
            } ]
          },
          "boundAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time",
            "description" : "When the system sources last moved to a repository or agent; null until configured"
          },
          "sources" : {
            "type" : "array",
            "description" : "Database first; empty until configured",
            "items" : {
              "$ref" : "#/components/schemas/SelfBackupSource"
            }
          }
        },
        "required" : [ "agentId", "boundAt", "configured", "repository", "sources" ]
      },
      "SelfBackupRepository" : {
        "type" : "object",
        "description" : "Where the self-backup goes, as the built-in agent's last Register describes it",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "backend" : {
            "type" : "string",
            "description" : "restic backend scheme: s3, sftp, local, ...; empty once Register no longer lists it"
          },
          "repositoryId" : {
            "type" : [ "string", "null" ],
            "description" : "restic repository id; null until the agent reads it"
          },
          "local" : {
            "type" : "boolean",
            "description" : "On the server's own machine: warn about it on every page"
          }
        },
        "required" : [ "backend", "local", "name", "repositoryId" ]
      },
      "SelfBackupSource" : {
        "type" : "object",
        "description" : "One of the self-backup's two system sources",
        "properties" : {
          "role" : {
            "$ref" : "#/components/schemas/SystemSourceRole"
          },
          "sourceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "agentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "repositoryName" : {
            "type" : "string"
          }
        },
        "required" : [ "agentId", "repositoryName", "role", "sourceId" ]
      },
...
      "RunStatus" : {
        "type" : "string",
        "enum" : [ "queued", "dispatched", "running", "succeeded", "failed", "cancelled" ]
      },
...
      "RunTrigger" : {
        "type" : "string",
        "enum" : [ "schedule", "manual", "verification", "catch_up" ]
      },
...
      "SelfBackupRunStarted" : {
        "type" : "object",
        "description" : "A run of one system source",
        "properties" : {
          "role" : {
            "$ref" : "#/components/schemas/SystemSourceRole"
          },
          "sourceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "runId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "started" : {
            "type" : "boolean",
            "description" : "False when the source already had this run active (D6): nothing new was queued"
          }
        },
        "required" : [ "role", "runId", "sourceId", "started" ]
      },
      "SelfBackupRuns" : {
        "type" : "object",
        "description" : "The runs \"back up now\" follows: one per system source, database first",
        "properties" : {
          "runs" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SelfBackupRunStarted"
            }
          }
        },
        "required" : [ "runs" ]
      },
...
      "RunPage" : {
        "type" : "object",
        "description" : "A page of runs",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RunSummary"
            }
          },
          "nextCursor" : {
            "type" : [ "string", "null" ],
            "description" : "Pass as cursor to get the next page; null on the last page"
          }
        },
        "required" : [ "items", "nextCursor" ]
      },
      "RunSummary" : {
        "type" : "object",
        "description" : "A run in the list",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "sourceId" : {
            "type" : "string",
            "format" : "uuid",
            "description" : "The source the run backs up"
          },
          "sourceName" : {
            "type" : "string",
            "description" : "The source's current name; for a deleted source, its name when it was deleted"
          },
          "sourceDeleted" : {
            "type" : "boolean",
            "description" : "The source was deleted; its runs and snapshots stay"
          },
          "agentId" : {
            "type" : "string",
            "format" : "uuid",
            "description" : "The source's agent"
          },
          "trigger" : {
            "$ref" : "#/components/schemas/RunTrigger"
          },
          "status" : {
            "$ref" : "#/components/schemas/RunStatus"
          },
          "message" : {
            "type" : [ "string", "null" ],
            "description" : "Why it failed, was rejected or lost; null on success"
          },
          "queuedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "startedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "finishedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          }
        },
        "required" : [ "agentId", "finishedAt", "id", "message", "queuedAt", "sourceDeleted", "sourceId", "sourceName", "startedAt", "status", "trigger" ]
      },
...
    },
    "securitySchemes" : {
      "session" : {
        "type" : "apiKey",
        "name" : "sard_session",
        "in" : "cookie"
      }
    }
  }
}